refactor: Restructure components
