+++ src/app/departures/card/favorite/departure-card.component.scss
