+++ src/app/departures/card/favorite/departure-card.component.spec.ts
