+++ src/app/departures/card/quick-search/quick-departure-search-card.component.spec.ts
