+++ src/app/departures/departures.module.ts
@@ -5,17 +5,17 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { DeparturesPageRoutingModule } from './departures-routing.module';
 import { TranslateModule } from '@ngx-translate/core';
-import { SharedComponentsModule } from '../shared-components/shared-components.module';
+import { SharedComponentsModule } from '../shared/components/shared-components.module';
 
 import { DeparturesPage } from './departures.page';
-import { DepartureCardComponent } from '../departure-card/departure-card.component';
-import { DepartureItemComponent } from '../departure-item/departure-item.component';
-import { MonitoredStationEditorComponent } from '../monitored-station-editor/monitored-station-editor.component';
-import { StationPickerComponent } from '../station-picker/station-picker.component';
-import { QuickDepartureSearchCardComponent } from '../quick-departure-search-card/quick-departure-search-card.component';
-import { DepartureListComponent } from '../departure-list/departure-list.component';
-import { TimerComponent } from '../timer/timer.component';
-import { DepartureSkeletonsComponent } from '../departure-skeletons/departure-skeletons.component';
+import { DepartureCardComponent } from './card/favorite/departure-card.component';
+import { DepartureItemComponent } from './item/departure-item.component';
+import { MonitoredStationEditorComponent } from '../shared/components/station/editor/monitored-station-editor.component';
+import { StationPickerComponent } from '../shared/components/station/picker/station-picker.component';
+import { QuickDepartureSearchCardComponent } from './card/quick-search/quick-departure-search-card.component';
+import { DepartureListComponent } from './list/departure-list.component';
+import { TimerComponent } from '../shared/components/timer/timer.component';
+import { DepartureSkeletonsComponent } from './item/skeleton/departure-skeletons.component';
 
 @NgModule({
   imports: [
