+++ src/app/departures/item/departure-item.component.html
