+++ src/app/departures/item/skeleton/departure-skeletons.component.scss
