+++ src/app/settings/about/about.component.scss
