+++ src/app/settings/about/about.component.spec.ts
