+++ src/app/settings/about/about.component.ts
