+++ src/app/settings/settings.module.ts
@@ -6,11 +6,11 @@ import { IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
 
 import { SettingsPageRoutingModule } from './settings-routing.module';
-import { SharedComponentsModule } from '../shared-components/shared-components.module';
+import { SharedComponentsModule } from '../shared/components/shared-components.module';
 
 import { SettingsPage } from './settings.page';
-import { LanguagePickerComponent } from '../language-picker/language-picker.component';
-import { AboutComponent } from '../about/about.component';
+import { LanguagePickerComponent } from './language-picker/language-picker.component';
+import { AboutComponent } from './about/about.component';
 
 @NgModule({
   imports: [
