+++ src/app/shared/components/shared-components.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 
-import { AppTitleComponent } from '../app-title/app-title.component';
+import { AppTitleComponent } from './app-title/app-title.component';
 
 @NgModule({
   declarations: [
