+++ src/app/shared/components/station/editor/monitored-station-editor.component.html
