+++ src/app/shared/components/station/editor/monitored-station-editor.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
 
-import { MonitoredStation } from '../shared/models/monitored-station.model';
+import { MonitoredStation } from '../../../models/monitored-station.model';
 
 @Component({
   selector: 'app-monitored-station-editor',
