+++ src/app/shared/components/timer/timer.component.spec.ts
