+++ src/app/shared/components/timer/timer.component.ts
